Add vue to the project and styled it
Create calculator using VUE:
Create Header, answer (which hold the answer being showed),deleteAnswer
(which deletes the answer),and array of numbers, array of operators,
equalAnswer(which is a button that will change the answer binding to
the desired cualculation entered.)

Methods:
Use Regex to find characters.
Add newAnswer = which will add the numbers to the answer binding.
pickOperator = This picks the operator and replaced the operator if
needs to be changed.
deleteA = This will delete the last character from answer binding and
the binding that separates the number after the operator.

Things to add:
A function that will return the answer to the answer binding.
A function that will make sure I can only add one operator to the
calculation.